requirements.txt
discord.py==1.3.2
PyNaCl==1.3.0
pandas
dnspython==1.16.0
async-timeout==3.0.1